--- lrage/gui/gui_utils/table.py
@@ -1,2 +1,0 @@
-def a():
-    return 1
