+++ run_gui.sh
@@ -1,4 +1,4 @@
 export OPENAI_API_KEY="write your own OpenAI API key"
 
-python lrage/gui/lrage_gui.py \
+python /data1/home/ohs/workspace/LRAGE/lrage/gui_new/lrage_gui.py \
 --port 3123
